dashboard/requirements.txt
mysql-connector-python
pandas
dash
plotly
python-dotenv